implementando CRUD Animal

+++ src/main/java/br/com/adoteumpet/repositories/AnimalRepository.java
@@ -15,7 +15,7 @@ public interface AnimalRepository extends JpaRepository<Animal, Long> {
 //	public List<Animal> findAllByONG(Integer idOng);
 	
 	@Query("Select obj from Animal obj WHERE Ong_id = ?1")
-	public List<Animal> findAllByONG(Integer idOng);
+	public List<Animal> findAllByONG(Long idOng);
 
 
 }
